+++ src/FoxIDs/Logic/Tracks/SessionLoginUpPartyLogic.cs
@@ -1,8 +1,10 @@
 ﻿using FoxIDs.Infrastructure;
 using FoxIDs.Models;
 using FoxIDs.Models.Config;
+using FoxIDs.Models.Logic;
 using FoxIDs.Models.Session;
 using FoxIDs.Repository;
+using ITfoxtec.Identity;
 using Microsoft.AspNetCore.Http;
 using System;
 using System.Collections.Generic;
@@ -18,28 +20,40 @@ namespace FoxIDs.Logic
         private readonly ITenantDataRepository tenantDataRepository;
         private readonly UpPartyCookieRepository<SessionLoginUpPartyCookie> sessionCookieRepository;
 
-        public SessionLoginUpPartyLogic(FoxIDsSettings settings, TelemetryScopedLogger logger, ITenantDataRepository tenantDataRepository, UpPartyCookieRepository<SessionLoginUpPartyCookie> sessionCookieRepository, IHttpContextAccessor httpContextAccessor) : base(settings, httpContextAccessor)
+        public SessionLoginUpPartyLogic(FoxIDsSettings settings, TelemetryScopedLogger logger, ITenantDataRepository tenantDataRepository, UpPartyCookieRepository<SessionLoginUpPartyCookie> sessionCookieRepository, TrackCookieRepository<SessionTrackCookie> sessionTrackCookieRepository, IHttpContextAccessor httpContextAccessor) : base(settings, sessionTrackCookieRepository, httpContextAccessor)
         {
             this.logger = logger;
             this.tenantDataRepository = tenantDataRepository;
             this.sessionCookieRepository = sessionCookieRepository;
         }
 
-        public async Task CreateSessionAsync(IUpParty loginUpParty, DownPartySessionLink newDownPartyLink, long authTime, IEnumerable<Claim> claims)
+        public async Task CreateOrUpdateMarkerSessionAsync(IUpParty loginUpParty, DownPartySessionLink downPartyLink)
         {
-            if(SessionEnabled(loginUpParty))
+            await AddOrUpdateSessionTrackAsync(loginUpParty, downPartyLink);
+
+            logger.ScopeTrace(() => $"Create marker session for authentication method, Route '{RouteBinding.Route}'.");
+            var session = new SessionLoginUpPartyCookie();
+            session.IsMarkerSession = true;
+            session.LastUpdated = session.CreateTime;
+            await sessionCookieRepository.SaveAsync(loginUpParty, session, null);
+        }
+
+        public async Task CreateSessionAsync(IUpParty upParty, long authTime, LoginUserIdentifier loginUserIdentifier, IEnumerable<Claim> claims)
+        {
+            if(SessionEnabled(upParty))
             {
                 logger.ScopeTrace(() => $"Create session, Route '{RouteBinding.Route}'.");
 
                 var session = new SessionLoginUpPartyCookie
                 {
-                    Claims = claims.ToClaimAndValues(),      
+                    Claims = claims.ToClaimAndValues(),
                 };
-                AddDownPartyLink(session, newDownPartyLink);
                 session.CreateTime = authTime;
                 session.LastUpdated = authTime;
-                await sessionCookieRepository.SaveAsync(loginUpParty, session, GetPersistentCookieExpires(loginUpParty, session.CreateTime));
-                logger.ScopeTrace(() => $"Session created, User id '{session.UserId}', Session id '{session.SessionId}'.", GetSessionScopeProperties(session));
+                SetLoginUserIdentifier(session, loginUserIdentifier);
+                await AddOrUpdateSessionTrackWithClaimsAsync(upParty, session.Claims);
+                await sessionCookieRepository.SaveAsync(upParty, session, GetPersistentCookieExpires(upParty, session.CreateTime));
+                logger.ScopeTrace(() => $"Session created, User id '{session.UserIdClaim}', Session id '{session.SessionIdClaim}'.", GetSessionScopeProperties(session));
             }
             else
             {
@@ -47,30 +61,43 @@ namespace FoxIDs.Logic
             }
         }
 
-        public async Task<bool> UpdateSessionAsync(IUpParty upParty, DownPartySessionLink newDownPartyLink, SessionLoginUpPartyCookie session, IEnumerable<Claim> claims = null)
+        public async Task<bool> UpdateSessionAsync(IUpParty upParty, SessionLoginUpPartyCookie session, LoginUserIdentifier loginUserIdentifier = null, IEnumerable<Claim> claims = null)
         {
             logger.ScopeTrace(() => $"Update session, Route '{RouteBinding.Route}'.");
 
             var sessionEnabled = SessionEnabled(upParty);
             var sessionValid = SessionValid(session, upParty);
 
-            if (sessionEnabled && sessionValid)
+            if (!session.IsMarkerSession && sessionEnabled && sessionValid)
             {
-                AddDownPartyLink(session, newDownPartyLink);
                 session.LastUpdated = DateTimeOffset.UtcNow.ToUnixTimeSeconds();
                 if (claims?.Count() > 0)
                 {
                     session.Claims = UpdateClaims(session, claims);
                 }
+                if (loginUserIdentifier != null)
+                {
+                    SetLoginUserIdentifier(session, loginUserIdentifier);
+                }
+                await AddOrUpdateSessionTrackWithClaimsAsync(upParty, session.Claims);
                 await sessionCookieRepository.SaveAsync(upParty, session, GetPersistentCookieExpires(upParty, session.CreateTime));
-                logger.ScopeTrace(() => $"Session updated, Session id '{session.SessionId}'.", GetSessionScopeProperties(session));
+                logger.ScopeTrace(() => $"Session updated, Session id '{session.SessionIdClaim}'.", GetSessionScopeProperties(session));
                 return true;
             }
 
             SetScopeProperty(session, includeSessionId: false);
             await sessionCookieRepository.DeleteAsync(upParty);
-            logger.ScopeTrace(() => $"Session deleted, Session id '{session.SessionId}'.");
+            logger.ScopeTrace(() => $"Session deleted, Session id '{session.SessionIdClaim}'.");
             return false;
+        }
+
+        private void SetLoginUserIdentifier(SessionLoginUpPartyCookie session, LoginUserIdentifier loginUserIdentifier)
+        {
+            session.UserId = loginUserIdentifier.UserId;
+            session.Email = loginUserIdentifier.Email;
+            session.Phone = loginUserIdentifier.Phone;
+            session.Username = loginUserIdentifier.Username;
+            session.UserIdentifier = loginUserIdentifier.UserIdentifier;
         }
 
         private IEnumerable<ClaimAndValues> UpdateClaims(SessionLoginUpPartyCookie session, IEnumerable<Claim> claims)
@@ -99,34 +126,40 @@ namespace FoxIDs.Logic
             return sessionClaims;
         }
 
-        public async Task<(SessionLoginUpPartyCookie, User)> GetAndUpdateSessionCheckUserAsync(LoginUpParty loginUpParty, DownPartySessionLink newDownPartyLink)
+        public async Task<(SessionLoginUpPartyCookie, User)> GetAndUpdateSessionCheckUserAsync(LoginUpParty loginUpParty)
         {
             logger.ScopeTrace(() => $"Get and update session and check user, Route '{RouteBinding.Route}'.");
 
             var session = await sessionCookieRepository.GetAsync(loginUpParty);
-            if (session != null)
+            if (session != null && session.Claims?.Count() > 0)
             {
                 var sessionEnabled = SessionEnabled(loginUpParty);
                 var sessionValid = SessionValid(session, loginUpParty);
 
-                logger.ScopeTrace(() => $"User id '{session.UserId}' session exists, Enabled '{sessionEnabled}', Valid '{sessionValid}', Session id '{session.SessionId}', Route '{RouteBinding.Route}'.");
+                logger.ScopeTrace(() => $"User id '{session.UserIdClaim}' session exists, Enabled '{sessionEnabled}', Valid '{sessionValid}', Session id '{session.SessionIdClaim}', Route '{RouteBinding.Route}'.");
                 if (sessionEnabled && sessionValid)
                 {
-                    var id = await User.IdFormatAsync(new User.IdKey { TenantName = RouteBinding.TenantName, TrackName = RouteBinding.TrackName, Email = session.Email });
+                    var email = session.Email;
+                    if (session.Email.IsNullOrEmpty() && session.UserIdentifier.IsNullOrEmpty() && session.UserId.IsNullOrEmpty())
+                    {
+                        // For backwards before version 1.15.0 - January 2025, can be deleted in about a year from now.
+                        email = session.EmailClaim;
+                    }
+                    var id = await User.IdFormatAsync(new User.IdKey { TenantName = RouteBinding.TenantName, TrackName = RouteBinding.TrackName, Email = email, UserIdentifier = session.UserIdentifier, UserId = session.UserId });
                     var user = await tenantDataRepository.GetAsync<User>(id, false);
                     if (user != null && !user.DisableAccount)
                     {
-                        AddDownPartyLink(session, newDownPartyLink);
+                        await AddOrUpdateSessionTrackWithClaimsAsync(loginUpParty, session.Claims);
                         session.LastUpdated = DateTimeOffset.UtcNow.ToUnixTimeSeconds();
                         await sessionCookieRepository.SaveAsync(loginUpParty, session, GetPersistentCookieExpires(loginUpParty, session.CreateTime));
-                        logger.ScopeTrace(() => $"Session updated, Session id '{session.SessionId}'.", GetSessionScopeProperties(session));
+                        logger.ScopeTrace(() => $"Session updated, Session id '{session.SessionIdClaim}'.", GetSessionScopeProperties(session));
                         return (session, user);
                     }
                 }
 
                 SetScopeProperty(session, includeSessionId: false);
                 await sessionCookieRepository.DeleteAsync(loginUpParty);
-                logger.ScopeTrace(() => $"Session deleted, Session id '{session.SessionId}'.");
+                logger.ScopeTrace(() => $"Session deleted, Session id '{session.SessionIdClaim}'.");
             }
             else
             {                
@@ -136,29 +169,29 @@ namespace FoxIDs.Logic
             return (null, null);
         }
 
-        public async Task<SessionLoginUpPartyCookie> GetAndUpdateExternalSessionAsync(ExternalLoginUpParty extLoginUpParty, DownPartySessionLink newDownPartyLink)
+        public async Task<SessionLoginUpPartyCookie> GetAndUpdateExternalSessionAsync(ExternalLoginUpParty extLoginUpParty)
         {
             logger.ScopeTrace(() => $"Get and update external session, Route '{RouteBinding.Route}'.");
 
             var session = await sessionCookieRepository.GetAsync(extLoginUpParty);
-            if (session != null)
+            if (session != null && session.Claims?.Count() > 0)
             {
                 var sessionEnabled = SessionEnabled(extLoginUpParty);
                 var sessionValid = SessionValid(session, extLoginUpParty);
 
-                logger.ScopeTrace(() => $"User id '{session.UserId}' session (external login) exists, Enabled '{sessionEnabled}', Valid '{sessionValid}', Session id '{session.SessionId}', Route '{RouteBinding.Route}'.");
+                logger.ScopeTrace(() => $"User id '{session.UserIdClaim}' session (external login) exists, Enabled '{sessionEnabled}', Valid '{sessionValid}', Session id '{session.SessionIdClaim}', Route '{RouteBinding.Route}'.");
                 if (sessionEnabled && sessionValid)
                 {
-                    AddDownPartyLink(session, newDownPartyLink);
+                    await AddOrUpdateSessionTrackWithClaimsAsync(extLoginUpParty, session.Claims);
                     session.LastUpdated = DateTimeOffset.UtcNow.ToUnixTimeSeconds();
                     await sessionCookieRepository.SaveAsync(extLoginUpParty, session, GetPersistentCookieExpires(extLoginUpParty, session.CreateTime));
-                    logger.ScopeTrace(() => $"Session (external login) updated, Session id '{session.SessionId}'.", GetSessionScopeProperties(session));
+                    logger.ScopeTrace(() => $"Session (external login) updated, Session id '{session.SessionIdClaim}'.", GetSessionScopeProperties(session));
                     return session;
                 }
 
                 SetScopeProperty(session, includeSessionId: false);
                 await sessionCookieRepository.DeleteAsync(extLoginUpParty);
-                logger.ScopeTrace(() => $"Session (external login) deleted, Session id '{session.SessionId}'.");
+                logger.ScopeTrace(() => $"Session (external login) deleted, Session id '{session.SessionIdClaim}'.");
             }
             else
             {
@@ -178,7 +211,7 @@ namespace FoxIDs.Logic
                 var sessionEnabled = SessionEnabled(upParty);
                 var sessionValid = SessionValid(session, upParty);
 
-                logger.ScopeTrace(() => $"User id '{session.UserId}' session exists, Enabled '{sessionEnabled}', Valid '{sessionValid}', Session id '{session.SessionId}', Route '{RouteBinding.Route}'.");
+                logger.ScopeTrace(() => $"User id '{session.UserIdClaim}' session exists, Enabled '{sessionEnabled}', Valid '{sessionValid}', Session id '{session.SessionIdClaim}', Route '{RouteBinding.Route}'.");
                 if (sessionEnabled && sessionValid)
                 {
                     SetScopeProperty(session);
@@ -187,7 +220,7 @@ namespace FoxIDs.Logic
 
                 SetScopeProperty(session, includeSessionId: false);
                 await sessionCookieRepository.DeleteAsync(upParty);
-                logger.ScopeTrace(() => $"Session deleted, Session id '{session.SessionId}'.");
+                logger.ScopeTrace(() => $"Session deleted, Session id '{session.SessionIdClaim}'.");
             }
             else
             {
@@ -204,7 +237,7 @@ namespace FoxIDs.Logic
             if (session != null)
             {
                 await sessionCookieRepository.DeleteAsync(upParty);
-                logger.ScopeTrace(() => $"Session deleted, Session id '{session.SessionId}'.");
+                logger.ScopeTrace(() => $"Session deleted, Session id '{session.SessionIdClaim}'.");
                 return session;
             }
             else
